+++ lib/caesura/src/core/caesura-core.scala
@@ -52,7 +52,7 @@ package dsvRedesignations:
   given capitalizedWords: DsvRedesignation = _.uncamel.map(_.capitalize).join(t" ")
   given lowerWords: DsvRedesignation = _.uncamel.map(_.lower).join(t" ")
 
-extension [encodable: Encodable in Row](value: encodable) def dsv: Row = encodable.encode(value)
+extension [encodable: Encodable in Dsv](value: encodable) def dsv: Dsv = encodable.encode(value)
 
-extension [encodable: Encodable in Row](value: Seq[encodable])
-  def dsv: Dsv = Dsv(value.to(Stream).map(encodable.encode(_)))
+extension [encodable: Encodable in Dsv](value: Seq[encodable])
+  def dsv: Sheet = Sheet(value.to(Stream).map(encodable.encode(_)))
